requirements.txt
python-decouple==3.5
Django==3.2.9
PyJWT==2.3.0
whitenoise==5.3.0
psycopg2-binary==2.9.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0